Crate polars_utils

Re-exports

pub use idx_vec::UnitVec;

pub use index::IdxSize;

pub use index::NullableIdxSize;

pub use pl_str::unique_column_name;

Modules

abs_diff

algebraic_ops

aliases

arc

arena

arg_min_max

array

async_utilsasync-utils

binary_search

bool

broadcast

cache

calc_morsel_split

cardinality_sketch

cell

Copy pasted from std::cell::SyncUnsafeCell can be removed once the feature stabilizes.

chunked_bytes_cursor

chunks

clmul

collection

compression

concat_vec

cpuid

error

file

fixedringbuffer

float

float16

floor_divmod

fmt

hashing

idx_map

idx_vec

index

io

itertools

kahan_sum

levenshtein

live_timer

macros

marked_usize

mem

min_max

mmapmmap

nulls

option

order_statistic_tree

This module implements an order statistic multiset, which is implemented as a weight-balanced tree (WBT). It is based on the weight-balanced tree based on the following papers:

parma

A fast concurrent hash map, with a focus on read performance.

pl_path

pl_serializeserde

Centralized Polars serialization entry.

pl_str

priority

python_convert_registrypython

python_functionpython

range

regex_cache

relaxed_cell

row_counter

scratch_vec

select

slice

slice_enum

small_bytes

sort

sparse_init_vec

sync

syssysinfo

tick_counter

total_ord

unique_id

vec

version

with_drop

Macros

cached_regex

format_list

format_list_container

format_list_container_truncated

format_list_truncated

format_pl_refstr

format_pl_smallstr

format_tuple

matches_any_order

no_call_const

unitvec

with_match_physical_numeric_polars_type