polars_utils::atomic
Struct SyncCounter

pub struct SyncCounter { /* private fields */ }

A utility to create a sharable counter This does not implement drop as the user needs to decide when to drop it. Which is likely the moment the last thread is finished.

Implementations

impl SyncCounter

pub fn new(value: usize) -> Self

pub unsafe fn manual_drop(&mut self)

Safety

This will deref the pointer and after this all autoderef will be invalid.