polars_utils
Module binary_search

Traits

ExponentialSearch

Functions

find_first_ge_index

Find the index of the first element of arr that is greater or equal to val. Assumes that arr is sorted.

find_first_gt_index

Find the index of the first element of arr that is greater than val. Assumes that arr is sorted.