polars_utils
Module hashing

Structs

BytesHash

Contains a byte slice and a precomputed hash for that string. During rehashes, we will rehash the hash instead of the string, that makes rehashing cheap and allows cache coherent small hash tables.

HashPartitioner

Traits

DirtyHash

Functions

_boost_hash_combine

folded_multiply

hash_to_partition