polars_utils::idx_vec

Struct UnitVec

pub struct UnitVec<T> { /* private fields */ }

A type logically equivalent to Vec<T>, but which does not do a memory allocation until at least two elements have been pushed, storing the first element inside the UnitVec directly.

Uses IdxSize internally to store lengths, will panic if trying to reserve for more elements.

Implementations

impl<T> UnitVec<T>

pub fn new() -> Self

pub fn is_inline(&self) -> bool

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn capacity(&self) -> usize

pub fn clear(&mut self)

pub fn push(&mut self, idx: T)

pub unsafe fn push_unchecked(&mut self, idx: T)

Safety

Caller must ensure that UnitVec has enough capacity.

pub fn pop(&mut self) -> Option<T>

pub fn reserve(&mut self, additional: usize)

pub fn with_capacity(capacity: usize) -> Self

pub fn iter(&self) -> Iter<'_, T>

pub fn iter_mut(&mut self) -> IterMut<'_, T>

pub fn as_slice(&self) -> &[T]

pub fn as_mut_slice(&mut self) -> &mut [T]

impl<T: Copy> UnitVec<T>

pub fn retain(&mut self, f: impl FnMut(T) -> bool)

impl<T: Clone> UnitVec<T>

pub fn from_slice(sl: &[T]) -> Self