polars_utils::mmap

Struct Mmap

pub struct Mmap { /* private fields */ }

Available on crate feature mmap only.

A handle to an immutable memory mapped buffer.

A Mmap may be backed by a file, or it can be anonymous map, backed by volatile memory. Use [MmapOptions] or map() to create a file-backed memory map. To create an immutable anonymous memory map, first create a mutable anonymous memory map, and then make it immutable with [MmapMut::make_read_only()].

A file backed Mmap is created by &File reference, and will remain valid even after the File is dropped. In other words, the Mmap handle is completely independent of the File used to create it. For consistency, on some platforms this is achieved by duplicating the underlying file handle. The memory will be unmapped when the Mmap handle is dropped.

Dereferencing and accessing the bytes of the buffer may result in page faults (e.g. swapping the mapped pages into physical memory) though the details of this are platform specific.

Mmap is Sync and Send.

See [MmapMut] for the mutable version.

Safety

All file-backed memory map constructors are marked unsafe because of the potential for Undefined Behavior (UB) using the map if the underlying file is subsequently modified, in or out of process. Applications must consider the risk and take appropriate precautions when using file-backed maps. Solutions such as file permissions, locks or process-private (e.g. unlinked) files exist but are platform specific and limited.

Example

use memmap2::MmapOptions;
use std::io::Write;
use std::fs::File;

let file = File::open("README.md")?;
let mmap = unsafe { MmapOptions::new().map(&file)? };
assert_eq!(b"# memmap2", &mmap[0..9]);

Implementations

impl Mmap

pub unsafe fn map<T>(file: T) -> Result<Mmap, Error>
where T: MmapAsRawDesc,

Creates a read-only memory map backed by a file.

This is equivalent to calling MmapOptions::new().map(file).

Safety

See the type-level docs for why this function is unsafe.

Errors

This method returns an error when the underlying system call fails, which can happen for a variety of reasons, such as when the file is not open with read permissions.

Example

use std::fs::File;
use std::io::Read;

use memmap2::Mmap;

let mut file = File::open("LICENSE-APACHE")?;

let mut contents = Vec::new();
file.read_to_end(&mut contents)?;

let mmap = unsafe { Mmap::map(&file)?  };

assert_eq!(&contents[..], &mmap[..]);

pub fn make_mut(self) -> Result<MmapMut, Error>

Transition the memory map to be writable.

If the memory map is file-backed, the file must have been opened with write permissions.

Errors

This method returns an error when the underlying system call fails, which can happen for a variety of reasons, such as when the file is not open with writable permissions.

Example

use memmap2::Mmap;
use std::ops::DerefMut;
use std::io::Write;

let file = /* file opened with write permissions */
let mmap = unsafe { Mmap::map(&file)? };
// ... use the read-only memory map ...
let mut mut_mmap = mmap.make_mut()?;
mut_mmap.deref_mut().write_all(b"hello, world!")?;

pub fn advise(&self, advice: Advice) -> Result<(), Error>

Advise OS how this memory map will be accessed.

Only supported on Unix.

See madvise() map page.

pub unsafe fn unchecked_advise( &self, advice: UncheckedAdvice, ) -> Result<(), Error>

Advise OS how this memory map will be accessed.

Used with the [unchecked flags][UncheckedAdvice]. Only supported on Unix.

See madvise() map page.

pub fn advise_range( &self, advice: Advice, offset: usize, len: usize, ) -> Result<(), Error>

Advise OS how this range of memory map will be accessed.

Only supported on Unix.

The offset and length must be in the bounds of the memory map.

See madvise() map page.

pub unsafe fn unchecked_advise_range( &self, advice: UncheckedAdvice, offset: usize, len: usize, ) -> Result<(), Error>

Advise OS how this range of memory map will be accessed.

Used with the [unchecked flags][UncheckedAdvice]. Only supported on Unix.

The offset and length must be in the bounds of the memory map.

See madvise() map page.

pub fn lock(&self) -> Result<(), Error>

Lock the whole memory map into RAM. Only supported on Unix.

See mlock() map page.

pub fn unlock(&self) -> Result<(), Error>

Unlock the whole memory map. Only supported on Unix.

See munlock() map page.

pub unsafe fn remap( &mut self, new_len: usize, options: RemapOptions, ) -> Result<(), Error>

Adjust the size of the memory mapping.

This will try to resize the memory mapping in place. If [RemapOptions::may_move] is specified it will move the mapping if it could not resize in place, otherwise it will error.

Only supported on Linux.

See the mremap(2) man page.

Safety

Resizing the memory mapping beyond the end of the mapped file will result in UB should you happen to access memory beyond the end of the file.