polars_utils::parma::raw

Trait Key

pub trait Key {
    // Required methods
    fn align() -> usize;
    fn size(&self) -> usize;
    unsafe fn init(&self, ptr: *mut u8);
    unsafe fn get<'a>(ptr: *const u8) -> &'a Self;
    unsafe fn drop_in_place(ptr: *mut u8);
}

A low-level trait for keys.

This is used to allow unsized types such as str for keys.

Required Methods

fn align() -> usize

The alignment necessary for the key. Must return a power of two.

fn size(&self) -> usize

The size of the key in bytes.

unsafe fn init(&self, ptr: *mut u8)

Initialize the key in the given memory location.

Safety

The memory location must satisfy the specified size and alignment.

unsafe fn get<'a>(ptr: *const u8) -> &'a Self

Get a reference to the key from the given memory location.

Safety

The pointer must be valid and initialized with Key::init.

unsafe fn drop_in_place(ptr: *mut u8)

Drop the key in place.

Safety

The pointer must be valid and initialized with Key::init.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.