polars_utils::pl_path

Struct PlPath

pub struct PlPath { /* private fields */ }

Path represented as a UTF-8 string.

Equality and ordering are based on the string value, which can be sensitive to duplicate separators. as_std_path() can be used to return a &std::path::Path for comparisons / API access.

Implementations

impl PlPath

pub fn as_str(&self) -> &str

pub fn as_bytes(&self) -> &[u8]

pub fn as_os_str(&self) -> &OsStr

pub fn as_std_path(&self) -> &Path

pub fn to_ref_path(&self) -> PlRefPath

pub fn scheme(&self) -> Option<CloudScheme>

pub fn has_scheme(&self) -> bool

Shorthand for self.scheme().is_some().

pub fn strip_scheme(&self) -> &str

Return a string with the scheme prefix removed (if any).

pub fn file_name(&self) -> Option<&OsStr>

pub fn extension(&self) -> Option<&str>

pub fn parent(&self) -> Option<&str>

pub fn sliced(&self, range: Range<usize>) -> &PlPath

Slices the path.

pub fn strip_scheme_split_authority(&self) -> Option<(&str, &str)>

Strips the scheme, then returns the authority component, and the remaining string after the authority component. This can be understood as extracting the bucket/prefix for cloud URIs.

E.g. https://user@host:port/dir/file?param=value

Authority: user@host:port

Remaining: /dir/file?param=value

Note, for local / file: URIs, the returned authority will be empty, and the remainder will be the full URI.

Returns

(authority, remaining).

pub fn authority_end_position(&self) -> usize

Returns 0 if self.scheme() is None. Otherwise, returns i such that &self.to_str()[..i] trims to the authority.

If there is no ‘/’, separator found, i will simply be the length of the string.

This is except if the scheme is FileNoHostname, where instead i will be “file:”.len()

If self has no CloudScheme, returns 0

pub fn to_absolute_path(&self) -> PolarsResult<PlRefPath>

pub fn join(&self, other: impl AsRef<str>) -> PlRefPath

pub fn normalize_windows_path(path_str: &str) -> Option<PlRefPath>

Converts backslashes to forward-slashes, and removes \\?\ prefix.