polars_utils
Module pl_serialize

Available on crate feature serde only.

Centralized Polars serialization entry.

Currently provides two serialization scheme’s.

Self-describing (and thus more forward compatible) activated with FC: true

Compact activated with FC: false

Structs

SerializeOptions

Mainly used to enable compression when serializing the final outer value. For intermediate serialization steps, the function in the module should be used instead.

Functions

deserialize_from_reader

deserialize_impl

deserialize_map_bytes

Potentially avoids copying memory compared to a naive Vec::<u8>::deserialize.

serialize_into_writer

serialize_to_bytes