Module polars_utils::sort

Functions

arg_sort_ascending

perfect_sort⚠Non-target_family="wasm"

This is a perfect sort particularly useful for an arg_sort of an arg_sort The second arg_sort sorts indices from 0 to len so can be just assigned to the new index location.