Module polars_utils::total_ord

Structs

TotalOrdWrap

Traits

ToTotalOrd

This elides creating a TotalOrdWrap for types that don’t need it.

TotalEq

Alternative trait for Eq. By consistently using this we can still be generic w.r.t Eq while getting a total ordering for floats.

TotalHash

Alternative trait for Hash. By consistently using this we can still be generic w.r.t Hash while being able to hash floats.

TotalOrd

Alternative trait for Ord. By consistently using this we can still be generic w.r.t Ord while getting a total ordering for floats.

Functions

canonical_f32

Converts an f32 into a canonical form, where -0 == 0 and all NaNs map to the same value.

canonical_f64

Converts an f64 into a canonical form, where -0 == 0 and all NaNs map to the same value.